Added an external ADC, such that I could get it to work with the T-BEAM, instead of having to rewrite the power management stuff for the esp32, as this is still a prototype. This ADC uses SPI.
Remembering this is just the code for the prototype, I also just created an SPI_SWC_Sensor outside of the sensor heirachy (which should be reworked anyways). It is basically a child of the SWC_analog_sensor - although the base class should just be SWC_sensor.

I have created a single test for the sensor, which does nothing but print the voltage.