1) FileManager.h: Added a method for saving with retries and delays.
2) FORMATTING_MACROS.h: Changed the size of the two value arrays, as they appeared to be messing with the ability of the esp32 to read and write.

Opted not to investigate. It appears to work 99% of the time, however, sometimes on startup it will not work at all and behave very weirdly.

The first few reads, it will disconnect the BLE device.

3) Main

Finished the skeleton of the loop.

Still cannot make it work with a dynamic time-getting, and I am now just writting a constant.

Need to figure out a way to get the timestamping to work.